Drone Defense
Now, drones not yet involved in attack will perform defensive drowning. There are still some improvements that need to be made with this, but I wanted this bot in for some scrimmages while I'm away from Battlecode for the next few hours so I could see if there is a proof-of-concept here.